Scaffold sink directory
