+++ datafusion-examples/examples/kafka_rideshare.rs
@@ -1,3 +1,20 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 #![allow(dead_code)]
 #![allow(unused_variables)]
 #![allow(unused_imports)]
@@ -147,10 +164,10 @@ async fn main() {
 
     // print_stream(&windowed_df).await;
 
-    use datafusion::dataframe::FileWriter;
-    use datafusion::dataframe::FranzSink;
+    use datafusion::franz_sinks::FileSink;
+    use datafusion::franz_sinks::FranzSink;
 
-    let writer = FileWriter::new("out.json").unwrap();
+    let writer = FileSink::new("/tmp/out.jsonl").unwrap();
     let file_writer = Box::new(writer) as Box<dyn FranzSink>;
 
     let _ = windowed_df.sink(file_writer).await;
